refactor: 모집 시작 페이지 리팩토링
