--- src/hooks/admin/ApplicationOpenPage/usePeopleAmount.ts
@@ -1,51 +1,0 @@
-import React from 'react';
-import { useAtom, useSetAtom } from 'jotai';
-import { openStepAtom, timeTemplateAtom, weeklyPeopleAtom } from 'pages/admin/ApplicationOpenPage';
-import { postOpenApplication } from 'apis/admin/application/open';
-import { useNavigate } from 'react-router-dom';
-import { convertPath } from 'apis/convertURI';
-import useErrorHandler from 'error/useErrorHandler';
-
-const usePeopleAmount = (startWeekDate: string) => {
-  // 전역 상태 가져오기 (주간 인원수 배열)
-  const [weeklyAmount, setWeeklyAmount] = useAtom(weeklyPeopleAtom);
-
-  // 이전 단계로 넘어가기
-  const setStep = useSetAtom(openStepAtom);
-  const goPrevHandler = () => {
-    setStep('setTime');
-  };
-
-  // 서버에 저장 (모집 시작 요청)
-
-  const [timeTemplate, setTimeTemplate] = useAtom(timeTemplateAtom);
-  const { apiErrorHandler } = useErrorHandler();
-  const navigate = useNavigate();
-  const submitStartHandler = () => {
-    postOpenApplication({
-      weeklyAmount: weeklyAmount,
-      timeTemplate: timeTemplate,
-      startWeekDate: startWeekDate,
-    })
-      .then((res) => {
-        navigate(convertPath('/'));
-
-        // 상태 초기화
-        setWeeklyAmount([[], [], [], [], [], [], []]);
-        setStep('setTime');
-        setTimeTemplate([]);
-      })
-      .catch((err) => {
-        apiErrorHandler(err);
-      });
-  };
-
-  return {
-    timeTemplate,
-    submitStartHandler,
-    goPrevHandler,
-    weeklyAmount,
-  };
-};
-
-export default usePeopleAmount;
